feat: adding the appropiate handler exceptions for the service and controllers

+++ everywhere-backend/src/main/java/com/everywhere/backend/service/impl/CategoriaServiceImpl.java
@@ -10,8 +10,7 @@ import com.everywhere.backend.service.CategoriaService;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.stereotype.Service;
-import java.util.List;
-import java.util.stream.Collectors;
+import java.util.List; 
 
 @Service
 @RequiredArgsConstructor
@@ -23,7 +22,7 @@ public class CategoriaServiceImpl implements CategoriaService {
 	@Override
 	public List<CategoriaResponseDto> findAll() {
 		return categoriaRepository.findAll()
-				.stream().map(categoriaMapper::toResponseDto).collect(Collectors.toList());
+				.stream().map(categoriaMapper::toResponseDto).toList();
 	}
 
 	@Override
@@ -35,21 +34,30 @@ public class CategoriaServiceImpl implements CategoriaService {
 
 	@Override
 	public CategoriaResponseDto create(CategoriaRequestDto categoriaRequestDto) {
-		Categoria categoria = categoriaMapper.toEntity(categoriaRequestDto);
-		Categoria saved = categoriaRepository.save(categoria);
-		return categoriaMapper.toResponseDto(saved);
+		if(categoriaRepository.existsByNombreIgnoreCase(categoriaRequestDto.getNombre()))
+			throw new RuntimeException("Ya existe una categoría con el nombre: " + categoriaRequestDto.getNombre());
+		Categoria categoria = categoriaMapper.toEntity(categoriaRequestDto); 
+		return categoriaMapper.toResponseDto(categoriaRepository.save(categoria));
 	}
 
 	@Override
 	public CategoriaResponseDto patch(int id, CategoriaRequestDto categoriaRequestDto) {
 		Categoria categoria = categoriaRepository.findById(id)
-				.orElseThrow(() -> new RuntimeException("Categoria no encontrada"));
+				.orElseThrow(() -> new RuntimeException("Categoria no encontrada con ID: " + id));
+
+		if (categoriaRequestDto.getNombre() != null && 
+			!categoriaRequestDto.getNombre().equalsIgnoreCase(categoria.getNombre()) &&
+			categoriaRepository.existsByNombreIgnoreCase(categoriaRequestDto.getNombre()))
+			throw new RuntimeException("Ya existe una categoría con el nombre: " + categoriaRequestDto.getNombre());
+
 		categoriaMapper.updateEntityFromDTO(categoriaRequestDto, categoria);
 		return categoriaMapper.toResponseDto(categoriaRepository.save(categoria));
 	}
 
 	@Override
 	public void delete(int id) {
+		if (!categoriaRepository.existsById(id))
+			throw new RuntimeException("Categoria no encontrada con ID: " + id);
 		categoriaRepository.deleteById(id);
 	}
 }
