+++ everywhere-backend/src/main/java/com/everywhere/backend/service/impl/CotizacionServiceImpl.java
@@ -12,8 +12,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -75,14 +73,15 @@ public class CotizacionServiceImpl implements CotizacionService {
     }
 
     @Override
-    public Optional<CotizacionResponseDto> findById(Integer id) {
-        return cotizacionRepository.findById(id).map(cotizacionMapper::toResponse);
+    public CotizacionResponseDto findById(Integer id) {
+        return cotizacionRepository.findById(id).map(cotizacionMapper::toResponse)
+            .orElseThrow(()-> new ResourceNotFoundException("Cotización no encontrada con ID: " + id));
     }
 
     @Override
     public List<CotizacionResponseDto> findAll() {
         return cotizacionRepository.findAll()
-                .stream().map(cotizacionMapper::toResponse).collect(Collectors.toList());
+                .stream().map(cotizacionMapper::toResponse).toList();
     }
 
     @Override
@@ -127,6 +126,8 @@ public class CotizacionServiceImpl implements CotizacionService {
 
     @Override
     public void delete(Integer id) {
+        if (!cotizacionRepository.existsById(id))
+            throw new ResourceNotFoundException("Cotización no encontrada con ID: " + id);
         cotizacionRepository.deleteById(id);
     }
 
@@ -143,15 +144,13 @@ public class CotizacionServiceImpl implements CotizacionService {
 
         CotizacionResponseDto cotizacionDTO = cotizacionMapper.toResponse(cotizacion);
         List<DetalleCotizacionResponseDto> detallesCompletos = detalleCotizacionService.findByCotizacionId(id);
-        List<DetalleCotizacionSimpleDTO> detallesSimples = detallesCompletos.stream()
-                .map(cotizacionMapper::toDetalleSimple).collect(Collectors.toList());
-
+        List<DetalleCotizacionSimpleDTO> detallesSimples = detallesCompletos.stream().map(cotizacionMapper::toDetalleSimple).toList();
         return cotizacionMapper.toResponseWithDetalles(cotizacionDTO, detallesSimples);
     }
 
     @Override
     public List<CotizacionResponseDto> findCotizacionesSinLiquidacion() {
         List<Cotizacion> cotizaciones = cotizacionRepository.findCotizacionesSinLiquidacion();
-        return cotizaciones.stream().map(cotizacionMapper::toResponse).collect(Collectors.toList());
+        return cotizaciones.stream().map(cotizacionMapper::toResponse).toList();
     }
 }
